pyproject.toml
[tool.poetry]
name = "aiimagebot"
version = "0.1.0"
description = ""
authors = ["PurplePalette"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
"discord.py" = "^2.1.0"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
novelai-api = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
